requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.9b0
bleach==4.1.0
build==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cmake==3.21.2
colorama==0.4.4
coloredlogs==15.0.1
contextlib2==21.6.0
cryptography==3.4.8
docutils==0.17.1
humanfriendly==10.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
jeepney==0.7.1
keyring==23.2.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pathspec==0.9.0
pep517==0.11.0
pkginfo==1.7.1
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
readme-renderer==29.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
schema==0.7.4
scipy==1.7.1
SecretStorage==3.3.1
sentient-traders @ git+ssh://git@github.com/1057405bcltd/sentient-traders.git@66f228735658c9927c6e4fac5f92f00c2644eb16
sentient-util @ git+ssh://git@github.com/1057405bcltd/sentient-util.git@872bda2f307d5182a1590f1c569b9c9bc290d9cf
six==1.16.0
toml==0.10.2
tomli==1.2.1
tqdm==4.62.3
traders==0.1.0
twine==3.4.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
zipp==3.5.0
